Split out user status functions
to avoid circular dep which has the weird effect of just breaking
jest's mocking.
