+++ apps/web/src/hooks/useUserStatus.ts
@@ -12,38 +12,9 @@ import { logger as rootLogger } from "matrix-js-sdk/src/logger";
 import { useMatrixClientContext } from "../contexts/MatrixClientContext";
 import { useTypedEventEmitter } from "./useEventEmitter";
 import { useFeatureEnabled } from "./useSettings";
+import { UserStatus, validateUserStatus } from "../utils/userStatus";
 
 const logger = rootLogger.getChild("useUserStatus");
-
-export interface UserStatus {
-    emoji: string;
-    text: string;
-}
-
-const MAX_STATUS_TEXT_BYTES = 256;
-
-export function userStatusTextWithinMaxLength(text: string): boolean {
-    const textEncoder = new TextEncoder();
-    return textEncoder.encode(text).length <= MAX_STATUS_TEXT_BYTES;
-}
-
-export function validateUserStatus(rawUserStatus: unknown): UserStatus | undefined {
-    if (typeof rawUserStatus !== "object" || rawUserStatus === null) {
-        return undefined;
-    }
-    if ("emoji" in rawUserStatus === false || typeof rawUserStatus.emoji !== "string" || !rawUserStatus.emoji) {
-        return undefined;
-    }
-    if ("text" in rawUserStatus === false || typeof rawUserStatus.text !== "string" || !rawUserStatus.text) {
-        return undefined;
-    }
-    return {
-        emoji: rawUserStatus.emoji,
-        text: userStatusTextWithinMaxLength(rawUserStatus.text)
-            ? rawUserStatus.text
-            : `${rawUserStatus.text.slice(0, MAX_STATUS_TEXT_BYTES)}…`,
-    };
-}
 
 /**
  * Hook to get the MSC4426 user status for a given user ID. Returns undefined if the feature is disabled,
